Refine API structure
