+++ client/index.html
@@ -79,12 +79,10 @@
         <script src="app/app.js"></script>
           <!-- injector:js -->
           <script src="app/constant.js"></script>
-          <script src="app/directives/analysis-chart/analysis-chart.directive.js"></script>
           <script src="app/directives/analysis_chart/analysis_chart.directive.js"></script>
           <script src="app/directives/appliance_record/appliance_record.directive.js"></script>
           <script src="app/directives/back_button/back_button.directive.js"></script>
           <script src="app/directives/my_appliance/my_appliance.directive.js"></script>
-          <script src="app/directives/realtime-chart/realtime-chart.directive.js"></script>
           <script src="app/directives/realtime_chart/realtime_chart.directive.js"></script>
           <script src="app/directives/setting_button/setting_button.directive.js"></script>
           <script src="app/directives/sign_in/sign_in.directive.js"></script>
