+++ package.json
@@ -3,20 +3,21 @@
   "version": "0.0.0",
   "main": "server/app.js",
   "dependencies": {
-    "express": "~4.0.0",
-    "morgan": "~1.0.0",
     "body-parser": "~1.5.0",
-    "method-override": "~1.0.0",
-    "serve-favicon": "~2.0.1",
-    "cookie-parser": "~1.0.1",
-    "express-session": "~1.0.2",
-    "errorhandler": "~1.0.0",
-    "compression": "~1.0.1",
-    "lodash": "~2.4.1",
-    "ejs": "~0.8.4",
-    "mysql": "^2.5.4",
     "composable-middleware": "^0.3.0",
-    "connect-mongo": "^0.4.1"
+    "compression": "~1.0.1",
+    "connect-mongo": "^0.4.1",
+    "cookie-parser": "~1.0.1",
+    "ejs": "~0.8.4",
+    "errorhandler": "~1.0.0",
+    "express": "~4.0.0",
+    "express-session": "~1.0.2",
+    "lodash": "~2.4.1",
+    "method-override": "~1.0.0",
+    "morgan": "~1.0.0",
+    "mysql": "^2.5.4",
+    "request": "^2.55.0",
+    "serve-favicon": "~2.0.1"
   },
   "devDependencies": {
     "grunt": "~0.4.4",
