--- server/api/thing/index.js
@@ -1,10 +1,0 @@
-'use strict';
-
-var express = require('express');
-var controller = require('./thing.controller');
-
-var router = express.Router();
-
-router.get('/', controller.index);
-
-module.exports = router;
