+++ server/routes.js
@@ -9,10 +9,11 @@ var errors = require('./components/errors');
 module.exports = function(app) {
 
   // Insert routes below
+  app.use('/api/scan', require('./api/scan'));
+  app.use('/api/devices', require('./api/device'));
   app.use('/api/appliances', require('./api/appliance'));
   app.use('/api/users', require('./api/user'));
-  app.use('/api/things', require('./api/thing'));
-  
+
   // All undefined asset or api routes should return a 404
   app.route('/:url(api|auth|components|app|bower_components|assets)/*')
    .get(errors[404]);
